Fixed gpu overlpa bug. Now pass in the correct cortical column number of rows and columns (height and width of the cortical columns).
We should think about removing passing in the neighbourhood step sizes as these should be calculated to have the correct sizes. If they aren't correct then the gpu overlap function will fail as it ecpects step size and cortical column size to match so that their is a cortical column synapse permanenece value per synapse.