Added additional check befroe a column is activated due to the surroudnign active cols being less then the desired local activaity then instead of activcating we also chekc the columsn in the neighbourhood list (and colInNeighboursLists) and find the ones that have not been processed yet (not inhibited or active) and check that this column has an overlap score large enough to warant activating it.
Thsi chekc prevents situations where mroe then the deisred local activiaty of columsn gets activated because of the order of columsn beeing porcessed and the fact the columns are not always in each otheres neighbourhood lists.

Thsi additional chekc is probablyt not needed for cases where we have centered and symetrical neighbourhoods we could probably disable it in these cases which should be the most common use case but for now I did not want these bugs intreoduced while testign on toy examnples.