WIP: Added argument to enable or disable addign adiditonal tie breaker values to the parallel inhibiton or serial sorted inhibition calculation. This works but I think we still have a bug in the inhibition code where the inhibition desired local activity is not being adhered to for larger inputs.
The reason for adding the optional add tiebreaker code is that usually the overlap calcualtor should make overlap vlaues unique but in test cases sometimes we don;t have this and we still want inhibition to be deterministic.

The parallel_add_tie_breaker is still a WIP since it uses its own task graph instead of adding itself to the main taskflwo graph.